requirements.txt
Flask==2.2.2
Werkzeug==2.3.7
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.3
flask-login==0.6.2
flask-sqlalchemy==3.0.3
sqlalchemy==1.4.46
flask-bcrypt==1.0.1
flask-wtf==1.1.1
pytubefix==4.0.0
youtube-transcript-api==0.5.0
openai>=1.0.0
reportlab==3.6.12
gunicorn==20.1.0
requests==2.28.2
python-dotenv==1.0.0
email-validator==2.0.0
psycopg2-binary==2.9.5
google-api-python-client